fix mistakes, feedback
