+++ src/earn/EarnPoolInfoScreen.tsx
@@ -3,7 +3,7 @@ import BigNumber from 'bignumber.js'
 import { Duration, intervalToDuration } from 'date-fns'
 import React, { useMemo, useRef, useState } from 'react'
 import { Trans, useTranslation } from 'react-i18next'
-import { LayoutChangeEvent, Platform, StyleSheet, Text, View, ViewStyle } from 'react-native'
+import { LayoutChangeEvent, Platform, StyleSheet, Text, View } from 'react-native'
 import Animated, { useAnimatedScrollHandler, useSharedValue } from 'react-native-reanimated'
 import { SafeAreaView, useSafeAreaInsets } from 'react-native-safe-area-context'
 import AppAnalytics from 'src/analytics/AppAnalytics'
@@ -13,6 +13,7 @@ import Button, { BtnSizes, BtnTypes } from 'src/components/Button'
 import { formatValueToDisplay } from 'src/components/TokenDisplay'
 import TokenIcon, { IconSize } from 'src/components/TokenIcon'
 import Touchable from 'src/components/Touchable'
+import { Card } from 'src/earn/Card'
 import InfoIcon from 'src/icons/InfoIcon'
 import OpenLinkIcon from 'src/icons/OpenLinkIcon'
 import { useDollarsToLocalAmount } from 'src/localCurrency/hooks'
@@ -107,22 +108,6 @@ function TokenIcons({
           showNetworkIcon={showNetworkIcon}
         />
       ))}
-    </View>
-  )
-}
-
-export function Card({
-  children,
-  testID,
-  cardStyle,
-}: {
-  children: React.ReactNode
-  testID: string
-  cardStyle?: ViewStyle
-}) {
-  return (
-    <View testID={testID} style={[styles.card, cardStyle]}>
-      {children}
     </View>
   )
 }
@@ -747,13 +732,6 @@ const styles = StyleSheet.create({
     flexDirection: 'row',
   },
   contentContainer: {
-    gap: Spacing.Regular16,
-  },
-  card: {
-    padding: Spacing.Regular16,
-    borderColor: Colors.gray2,
-    borderWidth: 1,
-    borderRadius: 12,
     gap: Spacing.Regular16,
   },
   cardLineContainer: {
